fixed a warning
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@19223 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/interface/ScrollBar.cpp
@@ -244,7 +244,7 @@ BScrollBar::Archive(BMessage *data, bool deep) const
 	err = data->AddInt32("_orient",(int32)fOrientation);
 	if (err != B_OK)
 		return err;
-	err = data->AddInt32("_prop",fProportion);
+	err = data->AddInt32("_prop",(int32)fProportion);
 	
 	return err;
 }
